task-a/text_processor/requirements.txt
functions-framework==3.*
google-cloud-storage
google-cloud-firestore